pyproject.toml
[tool.poetry]
name = "LLMs-on-terminal"
version = "0.1.0"
description = "Chat with state of the art LLMs from the terminal."
authors = ["bhancock_ai <brandon@brandonhancock.io>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.13"
langchain-openai = "^0.3.20"
python-dotenv = "^1.0.1"
langchain = "^0.3.25"
langchain-community = "^0.3.25"
langchain-anthropic = "^0.3.0"
langchain-google-genai = "^2.0.0"
langchain-google-firestore = "^0.4.0"
firestore = "^0.0.8"
chromadb = "^0.6.0"
tiktoken = "^0.8.0"
sentence-transformers = "^3.3.0"
bs4 = "^0.0.2"
firecrawl-py = "^1.0.0"
langchainhub = "^0.1.21"
wikipedia = "^1.4.0"
tavily-python = "^0.5.0"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']
